Optimizations in Comparison and Misc modules. Rewritten cmp function to handle
cases where it can easily be seen which one is larger without pokeing and calling
mpfr_cmp function. less, lesseq, greater, greatereq and equal are now implemented
with cmp. This saves some memory and improves speed in a lot of cases.

minD and maxD are also optimized to handle "obvious" cases without calling the
corresponding mpfr* functions. This also saves some time and speed in a lot of
cases.

neg in Arithmetic also received a speedup in cases where the requested precision
is the same as the precision of the operand (for example in negate in Num class).
There seems to be quite a speedup and reduced memory consumption when using
negate a lot.

darcs-hash:20081025162047-48752-7a804b25eab6c8cd6e04d11513f8b9938b5d72a9.gz